+++ src/utils/setupFilesUtils.js
@@ -1,5 +1,5 @@
 // External Dependencies
-import { basename, join, sep } from 'node:path'
+import { basename, dirname, join, sep } from 'node:path'
 
 // Internal Dependencies
 import {
@@ -9,7 +9,8 @@ import {
   THEME_INDEX_TEMPLATE_LIQUID_FILE
 } from '../config/constants.js'
 import { handleWatcherEvent } from './Watcher.js'
-import { copyFileAndCreatePath, getFileContents } from './fileUtils.js'
+import { getFileContents } from './fileUtils.js'
+import { installFile } from '../installers/collectionInstaller.js'
 
 const setupFolderCue = join(sep, SETUP_FOLDER_NAME, sep)
 const templatesFolderCue = join(sep, TEMPLATES_FOLDER_NAME, sep)
@@ -28,15 +29,17 @@ function getSetupRelativePath(filePath) {
  * Initial Installation Of Setup Files
  * @param {Component[]} components
  * @param {string} installFolder
+ * @param {string} copyrightText
  * @return {Promise<void[]>}
  */
-export async function installSetupFiles(components, installFolder) {
+export async function installSetupFiles(components, installFolder, copyrightText) {
   const copyPromises = []
   components.forEach((component) => {
     component.files.setupFiles.forEach((setupFile) => {
-      const setupFileRelativePath = getSetupRelativePath(setupFile)
+      const setupFileRelativePath = dirname(getSetupRelativePath(setupFile))
       const installPath = join(installFolder, setupFileRelativePath)
-      copyPromises.push(copyFileAndCreatePath(setupFile, installPath))
+
+      copyPromises.push(installFile(setupFile, installPath, copyrightText))
     })
   })
   return Promise.all(copyPromises)